Make classes named "String" invalid
Also adds a SStringType because we still need to represent the type "String"
somehow. The alternative would've been to add an internal "String" class
declaration to the classes namespace at the beginning of the analyze() call, but
I figured that this approach is cleaner. Also, it may be easier to extend to
support proper strings in the future.