Alternative performance fix for slow GVN without sacrificing readability (#160)
These changes provide an alternative fix for poor GVN performance due to excessive hash collisions, leading to catastrophic amounts of equality checks on const nodes.

Prior to this commit, nodes with no preds shared the same hash with all other nodes of the same kind, regardless of configuration. This was done with the intent to keep the code short and under the assumption that very very few nodes from the same kind, would share the same predecessors but were configured different. This turned out to be wrong for Const and Address nodes, since they don't have any predecessors.